Re-Introduce a trayicon UI object
For now it only has a single menu entry for the settings so I can test that
out. Also has the icon.

Everything done as a resource so deployment is somewhat easier (where
should images and uifiles go??). An eclipse launch for running pyrcc4 is
provided that should work on any system that has pyrcc4 in PATH.