Rework settings/monitors usage
No deletion at the moment, but at least now updating the monitor list
instead of re-creating and no storing of the data. There's little use in
storing all the state and storing the actual state instead of the color is
a little problematic with the json module.

Also avoids keeping the settings, storing redundant data and dropping the
extra conversion. The from/toDict functions are not needed anymore either,
but kept for now for the tests.