Handle errors and cancellation of the job-loading
If an error occurs we notify the user but do not remove the server as it
may be only a temporary issue. Also handle the user clicking on the cancel
button properly by removing the server from the list again.

Use the accept slot instead of the close one for the dialog so we can
actually differ between cancelling the operation and the dialog being
closed by one of the signals indicating the thread is done.