Introduce a 'lastState' member to jobs
I'll need this to limit when a notification is shown on the systray icon.
I don't want the notification to be shown if nothing happened to a job,
only if it starts to fail or has been fixed. So need a way to determine the
last state we saw for a job.