Don't bother refreshing server state when updating from settings
This can cause startup issues if there are problems connecting to one of
the servers. So better only setup the data objects and let the usual refresh
trigger the update.

Same goes with the ui-update after changing settings, not much point in
triggering that as there's going to be a run after the next timeout anyway.